requirements.txt
albumentations==1.4.8
comet_ml==3.43.1
coremltools==7.2
Flask==3.0.3
GitPython==3.1.43
GitPython==3.1.43
inference_sdk==0.12.1
ipython==8.12.3
langchain==0.2.5
matplotlib==3.9.0
mss==9.0.1
nncf==2.11.0
numpy==2.0.0
onnx==1.16.1
onnxruntime==1.17.1
onnxsim==0.4.36
openai==1.34.0
opencv_python==4.10.0.84
opencv_python_headless==4.8.0.74
openvino==2024.2.0
paddle==1.0.2
pafy==0.5.5
pandas==2.2.2
Pillow==10.3.0
psutil==5.9.8
pycocotools==2.0.8
PyYAML==6.0.1
PyYAML==6.0.1
Requests==2.32.3
roboflow==1.1.32
scipy==1.13.1
seaborn==0.13.2
setuptools==70.0.0
streamlit==1.33.0
tensorflow==2.16.1
tensorflowjs==4.20.0
tensorrt==10.1.0
tflite_runtime==2.14.0
tflite_support==0.4.4
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.2
tritonclient==2.46.0
ultralytics==8.2.35
ultralytics_thop==2.0.0
x2paddle==1.4.1
